refactor: form component
