--- src/components/inputs/index.ts
@@ -1,6 +1,0 @@
-export { default as TextInput } from './TextInput';
-export { default as CheckBoxInput } from './CheckBoxInput';
-export { default as DateInput } from './DateInput';
-export { default as FileInput } from './FileInput';
-export { default as RadioInput } from './RadioInput';
-export { default as SelectInput } from './SelectInput';
